sensors_recent: Improve gateway filtering
This allows filtering for multiple gateways (using comma-separated values in
the "gateways" parameter). The old "gateway" parameter is still processed for
compatibility. It also adds an show_other_gateways parameter, which can be used
to (not) show gateways other than the ones filtered. The default on new links
is changed to *do* show other gateways.