data: Fetch observations in a single query
Before, it would fetch observations first and then for each observation
do a new query to get the details about the observed species. This was
not as efficient as it could be, and because the outer query uses
MYSQLI_USE_RESULT, would cause a "Commands out of sync; you can't run
this command now" MySQL error when running the inner query.

By using a single query with joins, this is all fixed.

Note that in practice, the out of sync error did not seem to occur in
all cases, since this code has actually been working recently. However,
the new code should be better in all cases anyway.