health: Invalidate cache when script changes
Before, cache would only be invalidated when a new message arrived, but
that means that any offline nodes would be cached forever. This stores
the timestamp when the cache entry was created and compares that with
the modification date of the script.

The cache_timestamp field is added to the database as DEFAULT
CURRENT_TIMESTAMP() field, so the database manages it automatically.
Note that the `timestamp` field for measurements has incorrect timezones
(it has UTC stored as localtime, see #4), so the cache_timestamp and
last_seen fields are now in different timezones.

Additionally, this allows using ?nocache=1 to explicitly invalidate the
cache, and includes a `from_cache` field in the JSON response for easier
debugging.

Because the cache query now filters out rows, it might return 0 rows
even when a row is present for the current node. This means the old
behaviour of "INSERT when 0 rows, UPDATE otherwise" no longer works.
Instead, this now uses a REPLACE query, which automatically removes any
existing row with the same primary key, which works correctly and also
needs less code.

To keep the cache query simpler, it is also converted to a prepared
statement, which also potentially fixes some SQL escaping issues.