data: Fix invalid JSON for invalid coordinates
Since #14, only `null` values are omitted from the JSON output (rather
than all falsey values) to ensure that e.g. zero temperature is kept in
the JSON output. However, there was some code that used the empty string
to indicate an invalid position, which would now be put into the JSON
verbatim, making the JSON invalid.

This commit changes that code to use `null` for invalid positions
instead, which are now correctly omitted from the JSON. For CSV, this
does not matter, since `null` is converted to the empty string by PHP in
string concatenation automatically.